+++ chrome_target.go
@@ -104,6 +104,7 @@ type ChromeTarget struct {
 	Tracing              *gcdapi.Tracing
 	Tethering            *gcdapi.Tethering
 	Testing              *gcdapi.Testing
+	Fetch                *gcdapi.Fetch
 
 	Target      *TargetInfo              // The target information see, TargetInfo
 	sendCh      chan *gcdmessage.Message // The channel used for API components to send back to use
@@ -175,6 +176,7 @@ func (c *ChromeTarget) Init() {
 	// if stable channel you'll need to comment this out
 	c.Performance = gcdapi.NewPerformance(c)
 	c.Testing = gcdapi.NewTesting(c)
+	c.Fetch = gcdapi.NewFetch(c)
 }
 
 // clean up this target