+++ gcdapi/browser.go
@@ -311,8 +311,8 @@ func (c *Browser) GetWindowBounds(windowId int) (*BrowserBounds, error) {
 }
 
 type BrowserGetWindowForTargetParams struct {
-	// Devtools agent host id.
-	TargetId string `json:"targetId"`
+	// Devtools agent host id. If called as a part of the session, associated targetId is used.
+	TargetId string `json:"targetId,omitempty"`
 }
 
 // GetWindowForTargetWithParams - Get the browser window that contains the devtools target.
@@ -349,7 +349,7 @@ func (c *Browser) GetWindowForTargetWithParams(v *BrowserGetWindowForTargetParam
 }
 
 // GetWindowForTarget - Get the browser window that contains the devtools target.
-// targetId - Devtools agent host id.
+// targetId - Devtools agent host id. If called as a part of the session, associated targetId is used.
 // Returns -  windowId - Browser window id. bounds - Bounds information of the window. When window state is 'minimized', the restored window position and size are returned.
 func (c *Browser) GetWindowForTarget(targetId string) (int, *BrowserBounds, error) {
 	var v BrowserGetWindowForTargetParams
@@ -377,4 +377,26 @@ func (c *Browser) SetWindowBounds(windowId int, bounds *BrowserBounds) (*gcdmess
 	v.WindowId = windowId
 	v.Bounds = bounds
 	return c.SetWindowBoundsWithParams(&v)
+}
+
+type BrowserSetDockTileParams struct {
+	//
+	BadgeLabel string `json:"badgeLabel,omitempty"`
+	// Png encoded image.
+	Image string `json:"image,omitempty"`
+}
+
+// SetDockTileWithParams - Set dock tile details, platform-specific.
+func (c *Browser) SetDockTileWithParams(v *BrowserSetDockTileParams) (*gcdmessage.ChromeResponse, error) {
+	return gcdmessage.SendDefaultRequest(c.target, c.target.GetSendCh(), &gcdmessage.ParamRequest{Id: c.target.GetId(), Method: "Browser.setDockTile", Params: v})
+}
+
+// SetDockTile - Set dock tile details, platform-specific.
+// badgeLabel -
+// image - Png encoded image.
+func (c *Browser) SetDockTile(badgeLabel string, image string) (*gcdmessage.ChromeResponse, error) {
+	var v BrowserSetDockTileParams
+	v.BadgeLabel = badgeLabel
+	v.Image = image
+	return c.SetDockTileWithParams(&v)
 }